move to span-link
