+++ src/index.spec.ts
@@ -5,6 +5,7 @@
  * This product includes software developed at Datadog (https://www.datadoghq.com/).
  * Copyright 2021 Datadog, Inc.
  */
+
 jest.mock("./monitors.ts", () => {
   return {
     setMonitors: async (shouldThrow: Boolean) => {
@@ -16,18 +17,11 @@ jest.mock("./monitors.ts", () => {
   };
 });
 
-
-// tslint:disable-next-line:no-var-requires
-const stepFunctionsHelper = require("./step-functions-helper");
-stepFunctionsHelper.updateDefinitionString = jest.fn().mockImplementation();
 const ServerlessPlugin = require("./index");
 
-import {mergeStepFunctionAndLambdaTraces} from "./index";
 import mock from "mock-fs";
 import { FunctionDefinition } from "serverless";
 import Aws from "serverless/plugins/aws/provider/awsProvider";
-import Serverless from "serverless";
-import Service from "serverless/classes/Service";
 
 const SEM_VER_REGEX =
   /^v(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(-(0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(\.(0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*)?(\+[0-9a-zA-Z-]+(\.[0-9a-zA-Z-]+)*)?$/;
@@ -1876,96 +1870,6 @@ describe("ServerlessPlugin", () => {
       expect(serverless.service.stepFunctions.stateMachines.stepfunction1.loggingConfig.includeExecutionData).toEqual(
         true,
       );
-    });
-  });
-});
-
-describe("mergeStepFunctionAndLambdaTraces option related tests", () => {
-  function serviceWithResources(resources?: Record<string, any>, serviceName = "my-service"): Service {
-    const service = {
-      getServiceName: () => serviceName,
-      serverless: {
-        cli: {
-          log: () => "",
-        },
-      },
-      provider: {
-        name: "",
-        stage: "",
-        region: "",
-        versionFunctions: true,
-        compiledCloudFormationTemplate: {
-          Resources: resources as any,
-          Outputs: {},
-        },
-        logs: {
-          restApi: true,
-          httpApi: true,
-          websocket: true,
-        },
-      },
-    };
-    return service as any;
-  }
-  describe("test mergeStepFunctionAndLambdaTraces", () => {
-    it("have no state machine in the resources", async () => {
-      const resources = {
-        "a-lambda-resource": {
-          Type: "AWS::Lambda::Function",
-        },
-      };
-      const service = serviceWithResources();
-      const serverless: Serverless = service.serverless;
-      mergeStepFunctionAndLambdaTraces(resources, serverless);
-      expect(stepFunctionsHelper.updateDefinitionString).toBeCalledTimes(0);
-    });
-
-    it("have one state machine in the resources", async () => {
-      const resources = {
-        "unit-test-state-machine": {
-          Type: "AWS::StepFunctions::StateMachine",
-          Properties: {
-            DefinitionString: {
-              "Fn::Sub": ["real-definition-string", {}],
-            },
-          },
-        },
-        "another-resource": {
-          Type: "AWS::Lambda::Function",
-        },
-      };
-      const service = serviceWithResources();
-      const serverless: Serverless = service.serverless;
-      mergeStepFunctionAndLambdaTraces(resources, serverless);
-      expect(stepFunctionsHelper.updateDefinitionString).toBeCalledTimes(1);
-    });
-
-    it("have two state machine in the resources", async () => {
-      const resources = {
-        "unit-test-state-machine": {
-          Type: "AWS::StepFunctions::StateMachine",
-          Properties: {
-            DefinitionString: {
-              "Fn::Sub": ["real-definition-string", {}],
-            },
-          },
-        },
-        "unit-test-state-machine2": {
-          Type: "AWS::StepFunctions::StateMachine",
-          Properties: {
-            DefinitionString: {
-              "Fn::Sub": ["real-definition-string", {}],
-            },
-          },
-        },
-        "another-resource": {
-          Type: "AWS::Lambda::Function",
-        },
-      };
-      const service = serviceWithResources();
-      const serverless: Serverless = service.serverless;
-      mergeStepFunctionAndLambdaTraces(resources, serverless);
-      expect(stepFunctionsHelper.updateDefinitionString).toBeCalledTimes(2);
     });
   });
 });
