+++ src/index.ts
@@ -47,7 +47,7 @@ import { setMonitors } from "./monitors";
 import { addOutputLinks, printOutputs } from "./output";
 import { enableTracing, TracingMode } from "./tracing";
 import { redirectHandlers } from "./wrapper";
-import { GeneralResource, updateDefinitionString } from "./step-functions-helper";
+import { mergeStepFunctionAndLambdaTraces } from "./span-link";
 
 // Separate interface since DefinitelyTyped currently doesn't include tags or env
 export interface ExtendedFunctionDefinition extends FunctionDefinition {
@@ -572,20 +572,5 @@ function checkForMultipleApiKeys(config: Configuration) {
 
   if (multipleApiKeysMessage) {
     throw new Error(`${multipleApiKeysMessage} should not be set at the same time.`);
-  }
-}
-
-export function mergeStepFunctionAndLambdaTraces(
-  resources: { [key: string]: GeneralResource },
-  serverless: Serverless,
-) {
-  for (const resourceName in resources) {
-    if (resources.hasOwnProperty(resourceName)) {
-      const resourceObj: GeneralResource = resources[resourceName];
-      if (resourceObj.Type === "AWS::StepFunctions::StateMachine") {
-        const definitionString = resourceObj.Properties?.DefinitionString!;
-        updateDefinitionString(definitionString, serverless, resourceName);
-      }
-    }
   }
 }
